+++ init.rb
@@ -5,4 +5,4 @@ Tog::Plugins.settings :tog_mail, "messages.default_folders"   => "Inbox Outbox",
 
 Tog::Interface.sections(:member).add "Messages", "/member/messages"   
        
-Tog::Plugins.observers << :mail_user_observer
+#Tog::Plugins.observers << :mail_user_observer