+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 Edge
 ----
+* Create the default folders for every user on the migration.